# api/src/utils/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Verbify - OpenAPI 3.0
  description: |-
    Verbify API endpoints - Server based on the OpenAPI 3.0 specification.  
  version: 1.0.11
servers:
  - url: http://localhost:8080/api/
...
paths:
...
  /courses/populated:
    get:
      tags:
        - courses
      summary: Get all existing courses populated with all information
      description: Get all existing courses populated with information
      operationId: getCoursesPopulated
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Courses'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Courses'
        '400':
          description: Invalid request
        '404':
          description: Course not found
        '405':
          description: Validation exception
...
  /users:
    get:
      tags:
        - users
      summary: Get all existing users
      description: Get all existing users
      operationId: getusers
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Users'
        '400':
          description: Invalid request
        '404':
          description: user not found
        '405':
          description: Validation exception
      security:
        - api_key: []
    post:
      tags:
        - users
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Users'
          application/xml:
            schema:
              $ref: '#/components/schemas/Users'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Users'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
            application/xml:
              schema:
                $ref: '#/components/schemas/Users'
      security:
        - api_key: []
  /auth/signup:
    post:
      tags:
        - auth
      summary: Create a user with register method
      description: Create a user with register method
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Users'
          application/xml:
            schema:
              $ref: '#/components/schemas/Users'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Users'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
            application/xml:
              schema:
                $ref: '#/components/schemas/Users'
...
components:
  schemas:
    Courses:
      type: object
      properties:
        id:
          type: string
          format: objectId
          example: '6427c38d08b86ce760a1fa47'
        name:
          type: string
          example: 'Inglés básico'
        description:
          type: string
          example: 'Curso de inglés para estudiantes principiantes'
        imageURL:
          type: string
          example: 'cover.png'
        level:
          type: string
          example: beginner
        isPublic:
          type: boolean
          example: true
        tags: 
          type: string
        modules: 
          type: array
          xml:
            name: modules
            wrapped: true
          items:
            $ref: '#/components/schemas/Modules'
        isComplete:
          type: boolean
          example: false
      xml:
        name: course
    Modules:
      type: object
      properties:
        id:
          type: string
          format: objectId
          example: '6427c38d08b86ce760a1fa47'
        name:
          type: string
          example: 'Introduction'
        subtitle:
          type: string
          example: 'Inglés conversacional'
        isCompleted:
          type: boolean
        classes:
          type: array
          xml:
            name: classes
            wrapped: true
          items:
            $ref: '#/components/schemas/Classes'
      xml: 
        name: module
    Classes:
      type: object
      properties:
        id:
          type: string
          format: objectId
          example: '6427c38d08b86ce760a1fa47'
        name:
          type: string
          example: 'Verb To Be'
        content:
          type: string
          example: 'Clase del verbo to be, explicado y con parte practica'
        videoURL:
          type: string
          example: 'https://www.youtube.com/watch?v=a6o5tq9pV5I'
        isCompleted:
          type: boolean
          example: false
      xml:
        name: classes
    Users:
      type: object
      properties:
        id:
          type: string
          format: objectId
          example: '6427c38d08b86ce760a1fa47'
        email:
          type: string
          example: john@email.com
        passwordHash:
          type: string
          example: '12345'
        phone:
          type: string
          example: '+54 011 4789455'
        bio:
          type: string
          example: Ingeniero
        country: 
          type: string
          example: Colombia
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        isSuscribed:
          type: boolean
          example: true
...
  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: header